+++ zproject/urls.py
@@ -86,6 +86,7 @@ from zerver.views.message_flags import (
 from zerver.views.message_send import render_message_backend, send_message_backend, zcommand_backend
 from zerver.views.muted_users import mute_user, unmute_user
 from zerver.views.onboarding_steps import mark_onboarding_step_as_read
+from zerver.views.pinned_views import add_pinned_view, get_pinned_views, update_pinned_view_location
 from zerver.views.presence import (
     get_presence_backend,
     get_status_backend,
@@ -345,6 +346,9 @@ v1_api_and_json_patterns = [
     # saved_snippets -> zerver.views.saved_snippets
     rest_path("saved_snippets", GET=get_saved_snippets, POST=create_saved_snippet),
     rest_path("saved_snippets/<int:saved_snippet_id>", DELETE=delete_saved_snippet),
+    # pinned_views -> zerver.views.pinned_views
+    rest_path("pinned_views", GET=get_pinned_views, POST=add_pinned_view),
+    rest_path("pinned_views/<str:view_id>", PATCH=update_pinned_view_location),
     # New scheduled messages are created via send_message_backend.
     rest_path(
         "scheduled_messages", GET=fetch_scheduled_messages, POST=create_scheduled_message_backend